requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.5
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.1.0
